requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
Authlib==0.10
black==18.9b0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
ecdsa==0.13
flake8==3.6.0
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
pytest==4.0.2
python-jose==3.0.1
PyYAML==4.2b4
redis==3.0.1
requests==2.21.0
rsa==4.0
six==1.12.0
toml==0.10.0
urllib3==1.24.1
Werkzeug==0.14.1
